+++ Plutarch/Subtype.hs
@@ -2,10 +2,10 @@
 {-# LANGUAGE FunctionalDependencies #-}
 {-# LANGUAGE UndecidableInstances #-}
 
-module Plutarch.TryFrom (
-  PTryFrom (..),
+module Plutarch.Subtype (
+  PSubtype (..),
   HRecP (..),
-  ptryFrom,
+  pdowncast,
 ) where
 
 import Data.Proxy (Proxy (Proxy))
@@ -85,7 +85,7 @@ import Data.Kind (Type)
 
 import Data.Coerce (coerce)
 
------------------------ The class PTryFrom ----------------------------------------------
+----------------------- The class PSubtype ----------------------------------------------
 
 {- |
     This checks the data structure for validity.
@@ -95,27 +95,27 @@ import Data.Coerce (coerce)
     Be aware this might get really expensive, so only
     use it if you cannot establish trust otherwise
     (e.g. via only checking a part of your Data with
-    PTryFrom)
+    PSubtype)
     Laws:
-      - the operation `ptryFrom` mustn't change the representation of the underlying data
-      - the operation `ptryFrom` must always prove the integrity of the whole target type
+      - the operation `pdowncast` mustn't change the representation of the underlying data
+      - the operation `pdowncast` must always prove the integrity of the whole target type
         - example:
-          `ptryFrom PData (PAsData (PBuiltinList PData))` must only succeed if the underlying
+          `pdowncast PData (PAsData (PBuiltinList PData))` must only succeed if the underlying
           representation is a `BuiltinList` containing any `PData`
         - all conversions are fallible, this happens if the representation doesn't match
           the expected type.
-      - the operation `ptryFrom @b @a` proves equality between the less expressive `PType` `a` and
+      - the operation `pdowncast @b @a` proves equality between the less expressive `PType` `a` and
         the more expressive `PType` `b`, hence the first element of the resulting Tuple
         must always be wrapped in `PAsData` if the origin type was `PData` (see law 1)
       - the result type `b` must always be safe than the origin type `a`, i.e. it must carry
         more information
 -}
-class PTryFrom (a :: PType) (b :: PType) where
-  type PTryFromExcess a b :: PType
-  ptryFrom' :: forall s r. Term s a -> ((Term s b, Reduce (PTryFromExcess a b s)) -> Term s r) -> Term s r
+class PSubtype (a :: PType) (b :: PType) where
+  type PSubtypeExcess a b :: PType
+  pdowncast' :: forall s r. Term s a -> ((Term s b, Reduce (PSubtypeExcess a b s)) -> Term s r) -> Term s r
 
-ptryFrom :: forall b a s r. PTryFrom a b => Term s a -> ((Term s b, Reduce (PTryFromExcess a b s)) -> Term s r) -> Term s r
-ptryFrom = ptryFrom'
+pdowncast :: forall b a s r. PSubtype a b => Term s a -> ((Term s b, Reduce (PSubtypeExcess a b s)) -> Term s r) -> Term s r
+pdowncast = pdowncast'
 
 ----------------------- Reducible and Flip ----------------------------------------------
 
@@ -137,31 +137,31 @@ instance Reducible (HRecP as s) where type Reduce (HRecP as s) = HRecGeneric (HR
 
 ----------------------- PData instances -------------------------------------------------
 
-instance PTryFrom PData (PAsData PInteger) where
-  type PTryFromExcess PData (PAsData PInteger) = Flip Term PInteger
-  ptryFrom' opq = runTermCont $ do
+instance PSubtype PData (PAsData PInteger) where
+  type PSubtypeExcess PData (PAsData PInteger) = Flip Term PInteger
+  pdowncast' opq = runTermCont $ do
     ver <- tcont $ plet (pasInt # opq)
     pure $ (punsafeCoerce opq, ver)
 
-instance PTryFrom PData (PAsData PByteString) where
-  type PTryFromExcess PData (PAsData PByteString) = Flip Term PByteString
-  ptryFrom' opq = runTermCont $ do
+instance PSubtype PData (PAsData PByteString) where
+  type PSubtypeExcess PData (PAsData PByteString) = Flip Term PByteString
+  pdowncast' opq = runTermCont $ do
     ver <- tcont $ plet (pasByteStr # opq)
     pure $ (punsafeCoerce opq, ver)
 
 instance
-  ( PTryFrom PData (PAsData a)
-  , PTryFrom PData (PAsData b)
+  ( PSubtype PData (PAsData a)
+  , PSubtype PData (PAsData b)
   ) =>
-  PTryFrom PData (PAsData (PBuiltinMap a b))
+  PSubtype PData (PAsData (PBuiltinMap a b))
   where
-  type PTryFromExcess PData (PAsData (PBuiltinMap a b)) = Flip Term (PBuiltinMap a b)
-  ptryFrom' opq = runTermCont $ do
+  type PSubtypeExcess PData (PAsData (PBuiltinMap a b)) = Flip Term (PBuiltinMap a b)
+  pdowncast' opq = runTermCont $ do
     verMap <- tcont $ plet (pasMap # opq)
     let verifyPair :: Term _ (PBuiltinPair PData PData :--> PBuiltinPair (PAsData a) (PAsData b))
         verifyPair = plam $ \tup -> unTermCont $ do
-          (verfst, _) <- tcont $ ptryFrom @(PAsData a) $ pfstBuiltin # tup
-          (versnd, _) <- tcont $ ptryFrom @(PAsData b) $ psndBuiltin # tup
+          (verfst, _) <- tcont $ pdowncast @(PAsData a) $ pfstBuiltin # tup
+          (versnd, _) <- tcont $ pdowncast @(PAsData b) $ psndBuiltin # tup
           pure $ ppairDataBuiltin # verfst # versnd
     ver <- tcont $ plet $ pmap # verifyPair # verMap
     pure (punsafeCoerce opq, ver)
@@ -173,9 +173,9 @@ instance
 -}
 
 -- TODO: add the excess inner type list
-instance PTryFrom PData (PAsData (PBuiltinList PData)) where
-  type PTryFromExcess PData (PAsData (PBuiltinList PData)) = Flip Term (PBuiltinList PData)
-  ptryFrom' opq = runTermCont $ do
+instance PSubtype PData (PAsData (PBuiltinList PData)) where
+  type PSubtypeExcess PData (PAsData (PBuiltinList PData)) = Flip Term (PBuiltinList PData)
+  pdowncast' opq = runTermCont $ do
     ver <- tcont $ plet (pasList # opq)
     pure $ (punsafeCoerce opq, ver)
 
@@ -183,19 +183,19 @@ instance PTryFrom PData (PAsData (PBuiltinList PData)) where
     Recover a `PBuiltinList (PAsData a)`
 -}
 instance
-  ( PTryFrom PData (PAsData a)
+  ( PSubtype PData (PAsData a)
   , PIsData a
   ) =>
-  PTryFrom PData (PAsData (PBuiltinList (PAsData a)))
+  PSubtype PData (PAsData (PBuiltinList (PAsData a)))
   where
-  type PTryFromExcess PData (PAsData (PBuiltinList (PAsData a))) = Flip Term (PBuiltinList (PAsData a))
-  ptryFrom' opq = runTermCont $ do
+  type PSubtypeExcess PData (PAsData (PBuiltinList (PAsData a))) = Flip Term (PBuiltinList (PAsData a))
+  pdowncast' opq = runTermCont $ do
     let lst :: Term _ (PBuiltinList PData)
         lst = pasList # opq
         verify :: Term _ (PData :--> PAsData a)
         verify = plam $ \e ->
           unTermCont $ do
-            (wrapped, _) <- tcont $ ptryFrom @(PAsData a) $ e
+            (wrapped, _) <- tcont $ pdowncast @(PAsData a) $ e
             pure wrapped
     ver <- tcont $ plet $ pmap # verify # lst
     pure $ (punsafeCoerce opq, ver)
@@ -204,22 +204,22 @@ instance
     Recover a `PAsData (PBuiltinPair a b)`
 -}
 instance
-  ( PTryFrom PData a
+  ( PSubtype PData a
   , a ~ PAsData a'
   , PIsData a'
-  , PTryFrom PData b
+  , PSubtype PData b
   , b ~ PAsData b'
   , PIsData b'
   ) =>
-  PTryFrom PData (PAsData (PBuiltinPair a b))
+  PSubtype PData (PAsData (PBuiltinPair a b))
   where
-  type PTryFromExcess PData (PAsData (PBuiltinPair a b)) = Flip Term (PBuiltinPair a b)
-  ptryFrom' opq = runTermCont $ do
+  type PSubtypeExcess PData (PAsData (PBuiltinPair a b)) = Flip Term (PBuiltinPair a b)
+  pdowncast' opq = runTermCont $ do
     tup <- tcont $ plet (pfromData $ punsafeCoerce opq)
     let fst' :: Term _ a
-        fst' = unTermCont $ fst <$> tcont (ptryFrom @a $ pforgetData $ pfstBuiltin # tup)
+        fst' = unTermCont $ fst <$> tcont (pdowncast @a $ pforgetData $ pfstBuiltin # tup)
         snd' :: Term _ b
-        snd' = unTermCont $ fst <$> tcont (ptryFrom @b $ pforgetData $ psndBuiltin # tup)
+        snd' = unTermCont $ fst <$> tcont (pdowncast @b $ pforgetData $ psndBuiltin # tup)
     ver <- tcont $ plet $ ppairDataBuiltin # fst' # snd'
     pure $ (punsafeCoerce opq, ver)
 
@@ -227,15 +227,15 @@ instance
 
 -- We could have a more advanced instance but it's not needed really.
 
-newtype ExcessForField (a :: PType) (s :: S) = ExcessForField (Term s (PAsData a), Reduce (PTryFromExcess PData (PAsData a) s))
+newtype ExcessForField (a :: PType) (s :: S) = ExcessForField (Term s (PAsData a), Reduce (PSubtypeExcess PData (PAsData a) s))
 
-instance Reducible (PTryFromExcess PData (PAsData a) s) => Reducible (ExcessForField a s) where
-  type Reduce (ExcessForField a s) = (Term s (PAsData a), Reduce (PTryFromExcess PData (PAsData a) s))
+instance Reducible (PSubtypeExcess PData (PAsData a) s) => Reducible (ExcessForField a s) where
+  type Reduce (ExcessForField a s) = (Term s (PAsData a), Reduce (PSubtypeExcess PData (PAsData a) s))
 
 -- FIXME: Should we always succede? If we always succede, performance would increase a lot.
-instance PTryFrom (PBuiltinList PData) (PDataRecord '[]) where
-  type PTryFromExcess (PBuiltinList PData) (PDataRecord '[]) = HRecP '[]
-  ptryFrom' opq = runTermCont $ do
+instance PSubtype (PBuiltinList PData) (PDataRecord '[]) where
+  type PSubtypeExcess (PBuiltinList PData) (PDataRecord '[]) = HRecP '[]
+  pdowncast' opq = runTermCont $ do
     _ :: Term _ PUnit <-
       tcont . plet . pforce $
         pchooseListBuiltin # opq # pdelay (pcon PUnit) # pdelay (ptraceError "list is longer than zero")
@@ -245,23 +245,23 @@ type family UnHRecP (x :: PType) :: [(Symbol, PType)] where
   UnHRecP (HRecP as) = as
 
 instance
-  ( PTryFrom PData (PAsData pty)
-  , PTryFrom (PBuiltinList PData) (PDataRecord as)
-  , PTryFromExcess (PBuiltinList PData) (PDataRecord as) ~ HRecP ase
+  ( PSubtype PData (PAsData pty)
+  , PSubtype (PBuiltinList PData) (PDataRecord as)
+  , PSubtypeExcess (PBuiltinList PData) (PDataRecord as) ~ HRecP ase
   ) =>
-  PTryFrom (PBuiltinList PData) (PDataRecord ((name ':= pty) ': as))
+  PSubtype (PBuiltinList PData) (PDataRecord ((name ':= pty) ': as))
   where
   type
-    PTryFromExcess (PBuiltinList PData) (PDataRecord ((name ':= pty) ': as)) =
+    PSubtypeExcess (PBuiltinList PData) (PDataRecord ((name ':= pty) ': as)) =
       HRecP
         ( '(name, ExcessForField pty)
-            ': UnHRecP (PTryFromExcess (PBuiltinList PData) (PDataRecord as))
+            ': UnHRecP (PSubtypeExcess (PBuiltinList PData) (PDataRecord as))
         )
-  ptryFrom' opq = runTermCont $ do
+  pdowncast' opq = runTermCont $ do
     h <- tcont $ plet $ phead # opq
-    hv <- tcont $ ptryFrom @(PAsData pty) @PData h
+    hv <- tcont $ pdowncast @(PAsData pty) @PData h
     t <- tcont $ plet $ ptail # opq
-    tv <- tcont $ ptryFrom @(PDataRecord as) @(PBuiltinList PData) t
+    tv <- tcont $ pdowncast @(PDataRecord as) @(PBuiltinList PData) t
     pure (punsafeCoerce opq, HRecGeneric (HCons (Labeled hv) (coerce $ snd tv)))
 
 newtype Helper a b s = Helper (a s, b s)
@@ -270,22 +270,22 @@ instance (Reducible (a s), Reducible (b s)) => Reducible (Helper a b s) where
   type Reduce (Helper a b s) = (Reduce (a s), Reduce (b s))
 
 instance
-  ( PTryFrom (PBuiltinList PData) (PDataRecord as)
-  , PTryFromExcess (PBuiltinList PData) (PDataRecord as) ~ HRecP ase
+  ( PSubtype (PBuiltinList PData) (PDataRecord as)
+  , PSubtypeExcess (PBuiltinList PData) (PDataRecord as) ~ HRecP ase
   ) =>
-  PTryFrom PData (PAsData (PDataRecord as))
+  PSubtype PData (PAsData (PDataRecord as))
   where
   type
-    PTryFromExcess PData (PAsData (PDataRecord as)) =
-      Helper (Flip Term (PDataRecord as)) (PTryFromExcess (PBuiltinList PData) (PDataRecord as))
-  ptryFrom' opq = runTermCont $ do
-    l <- snd <$> (tcont $ ptryFrom @(PAsData (PBuiltinList PData)) opq)
-    r <- tcont $ ptryFrom @(PDataRecord as) l
+    PSubtypeExcess PData (PAsData (PDataRecord as)) =
+      Helper (Flip Term (PDataRecord as)) (PSubtypeExcess (PBuiltinList PData) (PDataRecord as))
+  pdowncast' opq = runTermCont $ do
+    l <- snd <$> (tcont $ pdowncast @(PAsData (PBuiltinList PData)) opq)
+    r <- tcont $ pdowncast @(PDataRecord as) l
     pure (punsafeCoerce opq, r)
 
-instance {-# OVERLAPPING #-} SumValidation 0 ys => PTryFrom PData (PAsData (PDataSum ys)) where
-  type PTryFromExcess PData (PAsData (PDataSum ys)) = Const ()
-  ptryFrom' opq = runTermCont $ do
+instance {-# OVERLAPPING #-} SumValidation 0 ys => PSubtype PData (PAsData (PDataSum ys)) where
+  type PSubtypeExcess PData (PAsData (PDataSum ys)) = Const ()
+  pdowncast' opq = runTermCont $ do
     x <- tcont $ plet $ pasConstr # opq
     constr <- tcont $ plet $ pfstBuiltin # x
     fields <- tcont $ plet $ psndBuiltin # x
@@ -298,7 +298,7 @@ class SumValidation (n :: Nat) (sum :: [[PLabeledType]]) where
 instance
   {-# OVERLAPPABLE #-}
   forall (n :: Nat) (x :: [PLabeledType]) (xs :: [[PLabeledType]]).
-  ( PTryFrom (PBuiltinList PData) (PDataRecord x)
+  ( PSubtype (PBuiltinList PData) (PDataRecord x)
   , SumValidation (n + 1) xs
   , KnownNat n
   ) =>
@@ -308,7 +308,7 @@ instance
     pif
       (fromInteger (natVal $ Proxy @n) #== constr)
       ( unTermCont $ do
-          _ <- tcont $ ptryFrom @(PDataRecord x) fields
+          _ <- tcont $ pdowncast @(PDataRecord x) fields
           pure $ popaque $ pcon PUnit
       )
       (validateSum @(n + 1) @xs constr fields)
@@ -322,47 +322,47 @@ instance {-# OVERLAPPING #-} SumValidation n '[] where
  unwrapped, then there is also the possibility to recover the whole thing but wrapped
 -}
 instance
-  ( PTryFrom a b
+  ( PSubtype a b
   , PIsData a
   , PIsData b
   ) =>
-  PTryFrom (PAsData a) (PAsData b)
+  PSubtype (PAsData a) (PAsData b)
   where
-  type PTryFromExcess (PAsData a) (PAsData b) = PTryFromExcess a b
-  ptryFrom' opq = runTermCont $ do
-    ver' <- snd <$> TermCont (ptryFrom @b @a (pfromData opq))
+  type PSubtypeExcess (PAsData a) (PAsData b) = PSubtypeExcess a b
+  pdowncast' opq = runTermCont $ do
+    ver' <- snd <$> TermCont (pdowncast @b @a (pfromData opq))
     pure $ (punsafeCoerce opq, ver')
 
-instance PTryFrom PData PData where
-  type PTryFromExcess PData PData = Const ()
-  ptryFrom' opq = runTermCont $ pure $ (opq, ())
+instance PSubtype PData PData where
+  type PSubtypeExcess PData PData = Const ()
+  pdowncast' opq = runTermCont $ pure $ (opq, ())
 
-instance PTryFrom PData (PAsData PData) where
-  type PTryFromExcess PData (PAsData PData) = Const ()
-  ptryFrom' opq = runTermCont $ pure (pdata opq, ())
+instance PSubtype PData (PAsData PData) where
+  type PSubtypeExcess PData (PAsData PData) = Const ()
+  pdowncast' opq = runTermCont $ pure (pdata opq, ())
 
 ----------------------- PIsDataReprInstances instance -----------------------------------
 
--- | you can instantiate `PTryFrom` for your own datatype as demonstrated in the spec
+-- | you can instantiate `PSubtype` for your own datatype as demonstrated in the spec
 instance
   ( PIsDataRepr a
   , SumValidation 0 (PIsDataReprRepr a)
   , PInner a b ~ PDataSum (PIsDataReprRepr a)
   ) =>
-  PTryFrom PData (PAsData (PIsDataReprInstances a))
+  PSubtype PData (PAsData (PIsDataReprInstances a))
   where
-  type PTryFromExcess PData (PAsData (PIsDataReprInstances a)) = Const ()
-  ptryFrom' opq = runTermCont $ do
+  type PSubtypeExcess PData (PAsData (PIsDataReprInstances a)) = Const ()
+  pdowncast' opq = runTermCont $ do
     let reprsum :: Term _ (PDataSum (PIsDataReprRepr a))
-        reprsum = pfromData $ unTermCont $ fst <$> TermCont (ptryFrom opq)
+        reprsum = pfromData $ unTermCont $ fst <$> TermCont (pdowncast opq)
     pure $ (pdata $ punsafeFrom reprsum, ())
 
 ----------------------- DerivePNewtype insatace -----------------------------------------
 
 instance
-  ( PTryFrom a b
+  ( PSubtype a b
   ) =>
-  PTryFrom a (DerivePNewtype c b)
+  PSubtype a (DerivePNewtype c b)
   where
-  type PTryFromExcess a (DerivePNewtype c b) = PTryFromExcess a b
-  ptryFrom' opq = runTermCont $ (\(inn, exc) -> (punsafeFrom inn, exc)) <$> tcont (ptryFrom @b @a opq)
+  type PSubtypeExcess a (DerivePNewtype c b) = PSubtypeExcess a b
+  pdowncast' opq = runTermCont $ (\(inn, exc) -> (punsafeFrom inn, exc)) <$> tcont (pdowncast @b @a opq)